V0.9: Remove crash for function def with zero args and warn on calls
Normal Verilog does not allow a function to be defined without at
least an input argument. This patch fixes a crash that was happening
when a function with no arguments was defined. It also adds checks
and prints a descriptive error message when calling a function without
an argument.

The development branch allows this in SV.